chore: Move some `tag_utils` code from `core` to `common`
