+++ core/common/src/lib.rs
@@ -3,5 +3,6 @@
 pub mod avm_string;
 pub mod buffer;
 pub mod sandbox;
+pub mod tag_utils;
 pub mod utils;
 pub mod xml;
