+++ scanner/src/execute.rs
@@ -6,7 +6,7 @@ use crate::logging::{ScanLogBackend, ThreadLocalScanLogger, LOCAL_LOGGER};
 use ruffle_core::backend::navigator::{NullExecutor, NullNavigatorBackend};
 use ruffle_core::limits::ExecutionLimit;
 use ruffle_core::swf::{decompress_swf, parse_swf};
-use ruffle_core::tag_utils::SwfMovie;
+use ruffle_core::tag_utils::movie_from_path;
 use ruffle_core::PlayerBuilder;
 use sha2::{Digest, Sha256};
 use std::io::{stdout, Write};
@@ -17,7 +17,7 @@ use std::time::{Duration, Instant};
 fn execute_swf(file: &Path) {
     let base_path = file.parent().unwrap();
     let executor = NullExecutor::new();
-    let movie = SwfMovie::from_path(file, None).unwrap();
+    let movie = movie_from_path(file, None).unwrap();
     let frame_time = 1000.0 / movie.frame_rate().to_f64();
     let player = PlayerBuilder::new()
         .with_log(ScanLogBackend::new())
